+++ src/main/java/com/lowdragmc/mbd2/integration/mekanism/trait/heat/CopiableHeatContainer.java
@@ -4,6 +4,7 @@ import com.lowdragmc.lowdraglib.syncdata.IContentChangeAware;
 import com.lowdragmc.lowdraglib.syncdata.ITagSerializable;
 import lombok.Getter;
 import lombok.Setter;
+import mekanism.api.heat.HeatAPI;
 import mekanism.api.heat.IHeatHandler;
 import net.minecraft.nbt.DoubleTag;
 import net.minecraft.nbt.Tag;
@@ -15,16 +16,17 @@ public class CopiableHeatContainer implements IHeatHandler, ITagSerializable<Tag
 
     public final double capacity;
     public final double inverseConduction;
-    public double temperature;
+    public double heat;
 
     public CopiableHeatContainer(double capacity, double inverseConduction) {
         this.capacity = capacity;
         this.inverseConduction = Math.max(1, inverseConduction);
+        this.heat = capacity * HeatAPI.AMBIENT_TEMP;
     }
 
     public CopiableHeatContainer copy() {
         CopiableHeatContainer copy = new CopiableHeatContainer(capacity, inverseConduction);
-        copy.temperature = temperature;
+        copy.heat = heat;
         return copy;
     }
 
@@ -35,7 +37,7 @@ public class CopiableHeatContainer implements IHeatHandler, ITagSerializable<Tag
 
     @Override
     public double getTemperature(int capacitor) {
-        return capacitor == 0 ? temperature : 0;
+        return capacitor == 0 ? heat / capacity : 0;
     }
 
     @Override
@@ -50,24 +52,21 @@ public class CopiableHeatContainer implements IHeatHandler, ITagSerializable<Tag
 
     @Override
     public void handleHeat(int capacitor, double transfer) {
-        if (capacitor == 0) {
-            var oldTemperature = temperature;
-            temperature = Math.min(capacity, temperature + transfer);
-            if (oldTemperature != temperature) {
-                onContentsChanged.run();
-            }
+        if (capacitor == 0 && transfer != 0) {
+            heat += transfer;
+            onContentsChanged.run();
         }
     }
 
     @Override
     public Tag serializeNBT() {
-        return DoubleTag.valueOf(temperature);
+        return DoubleTag.valueOf(heat);
     }
 
     @Override
     public void deserializeNBT(Tag nbt) {
         if (nbt instanceof DoubleTag tag) {
-            temperature = tag.getAsDouble();
+            heat = tag.getAsDouble();
         }
     }
 }