+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 # ChangeLog
 
 * Fixed Ingredient Candidates are linked between a base recipe and its copies
+* Refactor the mek heat to the temperature